+++ internal/s1ap/handover.go
@@ -13,7 +13,7 @@ var handoverContainerStub = s1ap.TransparentContainer{0x00}
 type HandoverRequiredParams struct {
 	MMEUES1APID       uint32
 	ENBUES1APID       uint32
-	CauseRadioNetwork int
+	CauseRadioNetwork int64
 
 	TargetMCC       string
 	TargetMNC       string
@@ -37,7 +37,7 @@ func BuildHandoverRequired(p HandoverRequiredParams) ([]byte, error) {
 		MMEUES1APID:  s1ap.MMEUES1APID(p.MMEUES1APID),
 		ENBUES1APID:  s1ap.ENBUES1APID(p.ENBUES1APID),
 		HandoverType: s1ap.HandoverTypeIntraLTE,
-		Cause:        s1ap.Cause{Group: s1ap.CauseGroupRadioNetwork, Value: p.CauseRadioNetwork},
+		Cause:        s1ap.Cause{Group: s1ap.CauseGroupRadioNetwork, Value: int(p.CauseRadioNetwork)},
 		TargetID: s1ap.TargetID{
 			TargeteNBID: s1ap.TargeteNBID{
 				GlobalENBID: s1ap.GlobalENBID{
@@ -86,7 +86,7 @@ func BuildHandoverRequestAcknowledge(p HandoverRequestAcknowledgeParams) ([]byte
 	for _, id := range p.FailedERABIDs {
 		failed = append(failed, s1ap.ERABItem{
 			ERABID: s1ap.ERABID(id),
-			Cause:  s1ap.Cause{Group: s1ap.CauseGroupRadioNetwork, Value: CauseRadioNetworkHOFailureInTarget},
+			Cause:  s1ap.Cause{Group: s1ap.CauseGroupRadioNetwork, Value: int(CauseRadioNetworkHOFailureInTarget)},
 		})
 	}
 
@@ -131,20 +131,20 @@ func BuildHandoverNotify(p HandoverNotifyParams) ([]byte, error) {
 	return m.Marshal()
 }
 
-func BuildHandoverCancel(mmeUES1APID, enbUES1APID uint32, cause int) ([]byte, error) {
+func BuildHandoverCancel(mmeUES1APID, enbUES1APID uint32, cause int64) ([]byte, error) {
 	m := &s1ap.HandoverCancel{
 		MMEUES1APID: s1ap.MMEUES1APID(mmeUES1APID),
 		ENBUES1APID: s1ap.ENBUES1APID(enbUES1APID),
-		Cause:       s1ap.Cause{Group: s1ap.CauseGroupRadioNetwork, Value: cause},
+		Cause:       s1ap.Cause{Group: s1ap.CauseGroupRadioNetwork, Value: int(cause)},
 	}
 
 	return m.Marshal()
 }
 
-func BuildHandoverFailure(mmeUES1APID uint32, cause int) ([]byte, error) {
+func BuildHandoverFailure(mmeUES1APID uint32, cause int64) ([]byte, error) {
 	m := &s1ap.HandoverFailure{
 		MMEUES1APID: s1ap.MMEUES1APID(mmeUES1APID),
-		Cause:       s1ap.Cause{Group: s1ap.CauseGroupRadioNetwork, Value: cause},
+		Cause:       s1ap.Cause{Group: s1ap.CauseGroupRadioNetwork, Value: int(cause)},
 	}
 
 	return m.Marshal()
